+++ App/discogs/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+discogs_bp = Blueprint('discogs', __name__)
+
+from . import routes
