+++ App/extensions.py
@@ -1,0 +1,8 @@
+import os
+
+from flask import current_app
+import redis
+
+REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")
+
+redis_client = redis.from_url(REDIS_URL)
