+++ App/spotify/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+spotify_bp = Blueprint('spotify', __name__)
+
+from . import routes
